backend/requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-CORS==4.0.0
PyJWT==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0
pymongo==4.4.1
gunicorn==21.2.0  # Important for production
Werkzeug==2.3.7
